now on a 12x12 grid. Much less initial epsilon for faster learning. Cleaned init files, made paths relative throughout, assuming only that micropolis-4bots is in the same directory as this.

+++ gym_micropolis/envs/tilemap.py
@@ -37,7 +37,7 @@ class TileMap(object):
             elif z_size and z_size > 1:
                 self.zoneSquares[z] = self.makeZoneSquare(self.zoneSize[z], self.zoneInts[z])
         # first dimension is one-hot zone encoding followed by zone int
-        self.zoneMap = np.zeros((self.num_zones + 1,micro.MAP_X, micro.MAP_Y), dtype=int)
+        self.zoneMap = np.zeros((self.num_zones + 1,micro.MAP_X, micro.MAP_Y), dtype=np.int8)
         self.zoneMap[-1,:,:] = self.zoneInts['Clear']
         self.zoneMap[self.zoneInts['Clear'],:,:] = 1
         self.micro = micro
